requirements.txt
tensorflow>=2.2.0
numpy
matplotlib
mlflow
scipy
scikit-image
click
keras-tuner
scikit-learn
requests